Move task checker and scheduler to its own cog
